New Motion
I've made a new motion script and tested it in the MotionTest Scene,
the main scene has not been changed yet. I still need to move the
Respawn() to a new script and write up some new functions to determine
where the racer is on the track. So, for now keep using the old
MotionController Script.
Also, the Player's machine gun will listen for the fire button event.
Let me know if it works!